Add cppcheck static-analysis gate + fix warnings it flagged
- .github/workflows/host-tests.yml (renamed "Host Tests & Static Analysis"):
  new job runs cppcheck --enable=warning over the full feature set with
  --error-exitcode=1. Uses the runner's system cppcheck, so it needs no
  PlatformIO registry (works independently of the firmware build).
- Fixes for every warning it reported:
  - data_logger: %lu logged a uint32_t timestamp -> cast to unsigned long.
  - Fully initialize the ATAK (multicastPort), CRSF/MAVLink (rxBuffer) and
    Acoustic (goertzelFilters/audioBuffer) constructors.
  - GPSModule: deleted copy ctor/operator= (it owns a raw HardwareSerial*).

Verified locally: full-feature `cppcheck --enable=warning` exits clean, and the
host parser suite still passes 27/27.

+++ src/data_logger.cpp
@@ -114,7 +114,7 @@ bool DataLogger::logDetection(const DetectionLogEntry& entry) {
     char logLine[256];
     snprintf(logLine, sizeof(logLine), 
              "[%lu] DETECTION | Module: %s | RSSI: %d dBm | Freq: %.2f MHz | Protocol: %s | GPS: %.6f,%.6f @ %.1fm\n",
-             entry.timestamp, entry.moduleName, entry.rssi, entry.frequency, entry.protocol,
+             (unsigned long)entry.timestamp, entry.moduleName, entry.rssi, entry.frequency, entry.protocol,
              entry.latitude, entry.longitude, entry.altitude);
     
     size_t written = logFile.print(logLine);
